fix review issues.

+++ frameworks/src/core/components/image_animator_component.cpp
@@ -214,7 +214,7 @@ bool ImageAnimatorComponent::ParseToFrames(jerry_value_t value)
         HILOG_ERROR(HILOG_MODULE_ACE, "fail to set images cause by out of memory.");
         return false;
     }
-    for (uint16_t idx = 0; idx < framesSize_; ++idx) {
+    for (uint8_t idx = 0; idx < framesSize_; ++idx) {
         jerry_value_t image = jerry_get_property_by_index(value, idx);
         if (jerry_value_is_undefined(image) || !jerry_value_is_object(image)) {
             HILOG_WARN(HILOG_MODULE_ACE, "the %{public}d frame is null or not an object.", idx);
